# api/storage/api-docs/api-docs.json
{"swagger":"2.0","info":{"title":"Talent documentations","license":{"name":"MIT"},"version":"1.0.0"},"host":"","basePath":"\/","schemes":["http"],"consumes":["application\/json"],"produces":["application\/json"],"paths":{"\/admin\/candidates":{"get":{"tags":["ADMIN:: Candidate's methods"],"summary":"Get list of candidates with pagination","description":"Get list of candidates with pagination","produces":["application\/json"],"parameters":[{"name":"page","in":"query","description":"Number of page","required":true,"type":"integer","default":"1"}],"responses":{"200":{"description":"User's data"}},"security":[{"Bearer":[]}]},"post":{"tags":["ADMIN:: Candidate's methods"],"summary":"Add new candidate","description":"","consumes":["multipart\/form-data"],"produces":["application\/json"],"parameters":[{"name":"resume","in":"formData","description":"Resume file","required":false,"type":"file"},{"name":"name","in":"formData","description":"Candidate's name","required":true,"type":"string"},{"name":"email","in":"formData","description":"Candidate's email","required":true,"type":"string"},{"name":"phone","in":"formData","description":"Candidate's phone","required":true,"type":"string"}],"responses":{"200":{"description":"User's data"}},"security":[{"Bearer":[]}]}},"\/admin\/candidates\/{id}":{"post":{"tags":["ADMIN:: Candidate's methods"],"summary":"Edit candidate","description":"","consumes":["multipart\/form-data"],"produces":["application\/json"],"parameters":[{"name":"id","in":"path","description":"Candidate's id","required":true,"type":"integer"},{"name":"new_resume","in":"formData","description":"Resume file","required":false,"type":"file"},{"name":"name","in":"formData","description":"Candidate's name","required":true,"type":"string"},{"name":"email","in":"formData","description":"Candidate's email","required":true,"type":"string"},{"name":"phone","in":"formData","description":"Candidate's phone","required":true,"type":"string"},{"name":"resume","in":"formData","description":"Candidate's link to resume (old)","required":true,"type":"string"}],"responses":{"200":{"description":"User's data"}},"security":[{"Bearer":[]}]}},"\/admin\/jobs":{"get":{"tags":["ADMIN:: Job's methods"],"summary":"Get list of jobs with pagination","description":"Get list of jobs with pagination","produces":["application\/json"],"parameters":[{"name":"page","in":"query","description":"Number of page","required":true,"type":"integer","default":"1"}],"responses":{"200":{"description":"Jobs data"}},"security":[{"Bearer":[]}]},"post":{"tags":["ADMIN:: Job's methods"],"summary":"Add new job","description":"Add new job","produces":["application\/json"],"parameters":[{"name":"data","in":"body","description":"Job's data","required":true,"schema":{"properties":{"client_id":{"type":"integer","default":""},"description":{"type":"string","default":""},"active":{"type":"integer","default":"0"},"completed":{"type":"integer","default":"0"},"assigned":{"description":"Talent ids; nullable","type":"integer","default":""}},"type":"object"}}],"responses":{"200":{"description":"Job's data"}},"security":[{"Bearer":[]}]}},"\/admin\/jobs\/{id}":{"put":{"tags":["ADMIN:: Job's methods"],"summary":"Edit the job","description":"Edit the job","produces":["application\/json"],"parameters":[{"name":"jobId","in":"path","description":"Number of job","required":true,"type":"integer","default":""},{"name":"data","in":"body","description":"Job's data","required":true,"schema":{"properties":{"client_id":{"type":"integer","default":""},"description":{"type":"string","default":""},"active":{"type":"integer","default":"0"},"completed":{"type":"integer","default":"0"},"assigned":{"description":"Talent ids; nullable","type":"integer","default":""}},"type":"object"}}],"responses":{"200":{"description":"Job's data"}},"security":[{"Bearer":[]}]}},"\/admin\/partner-candidates":{"get":{"tags":["ADMIN:: Partner-candidate's methods"],"summary":"Get list of partner-candidates with pagination","description":"Get list of partner-candidates with pagination","produces":["application\/json"],"parameters":[{"name":"page","in":"query","description":"Number of page","required":true,"type":"integer","default":"1"}],"responses":{"200":{"description":"User's data"}},"security":[{"Bearer":[]}]},"post":{"tags":["ADMIN:: Partner-candidate's methods"],"summary":"Add new partner-candidate","description":"","consumes":["multipart\/form-data"],"produces":["application\/json"],"parameters":[{"name":"resume","in":"formData","description":"Resume file","required":false,"type":"file"},{"name":"name","in":"formData","description":"Candidate's name","required":true,"type":"string"},{"name":"email","in":"formData","description":"Candidate's email","required":true,"type":"string"},{"name":"phone","in":"formData","description":"Candidate's phone","required":true,"type":"string"}],"responses":{"200":{"description":"User's data"}},"security":[{"Bearer":[]}]}},"\/admin\/partner-candidates\/{id}":{"post":{"tags":["ADMIN:: Partner-candidate's methods"],"summary":"Edit partner-candidate","description":"","consumes":["multipart\/form-data"],"produces":["application\/json"],"parameters":[{"name":"id","in":"path","description":"Candidate's id","required":true,"type":"integer"},{"name":"new_resume","in":"formData","description":"Resume file","required":false,"type":"file"},{"name":"name","in":"formData","description":"Candidate's name","required":true,"type":"string"},{"name":"email","in":"formData","description":"Candidate's email","required":true,"type":"string"},{"name":"phone","in":"formData","description":"Candidate's phone","required":true,"type":"string"},{"name":"resume","in":"formData","description":"Candidate's link to resume (old)","required":true,"type":"string"}],"responses":{"200":{"description":"Candidate's data"}},"security":[{"Bearer":[]}]}},"\/admin\/reviews":{"get":{"tags":["ADMIN:: Reviews methods"],"summary":"Get list of review with pagination","description":"Get list of review with pagination","produces":["application\/json"],"parameters":[{"name":"page","in":"query","required":true,"type":"integer","default":"1"}],"responses":{"200":{"description":"Reviews array"}},"security":[{"Bearer":[]}]}},"\/admin\/users":{"get":{"tags":["ADMIN:: User's methods"],"summary":"","description":"","produces":["application\/json"],"responses":{"200":{"description":"User's data"}},"security":[{"Bearer":[]}]},"post":{"tags":["ADMIN:: User's methods"],"summary":"Add new user","description":"","produces":["application\/json"],"parameters":[{"name":"data","in":"body","description":"User's data","required":true,"schema":{"properties":{"name":{"type":"string","default":""},"email":{"type":"string","default":""},"phone":{"type":"string","default":""},"user_role":{"properties":{"role_id":{"type":"integer","default":""}},"type":"object"}},"type":"object"}}],"responses":{"200":{"description":"User's data"}},"security":[{"Bearer":[]}]}},"\/admin\/users\/list":{"get":{"tags":["ADMIN:: User's methods"],"summary":"","description":"","produces":["application\/json"],"parameters":[{"name":"role","in":"query","description":"User's role","required":true,"type":"string"}],"responses":{"200":{"description":"User's data"}},"security":[{"Bearer":[]}]}},"\/admin\/users\/{id}":{"put":{"tags":["ADMIN:: User's methods"],"summary":"Edit user","description":"","produces":["application\/json"],"parameters":[{"name":"id","in":"path","description":"User's id","required":true,"type":"integer"},{"name":"data","in":"body","description":"User's data","required":true,"schema":{"properties":{"name":{"type":"string","default":""},"email":{"type":"string","default":""},"phone":{"type":"string","default":""},"user_role":{"properties":{"role_id":{"type":"integer","default":""}},"type":"object"}},"type":"object"}}],"responses":{"200":{"description":"User's data"}},"security":[{"Bearer":[]}]}},"\/sign-in":{"post":{"tags":["Authentications methods"],"summary":"Sign in user in the system","description":"Sign in user in the system","produces":["application\/json"],"parameters":[{"name":"data","in":"body","description":"Data for sign in","required":true,"schema":{"properties":{"email":{"type":"string","default":""},"password":{"type":"string","default":""}},"type":"object"}}],"responses":{"200":{"description":"User's data"}}}},"\/refresh-token":{"get":{"tags":["Authentications methods"],"summary":"","description":"","produces":["application\/json"],"responses":{"200":{"description":"User's data"}},"security":[{"Bearer":[]}]}},"\/sign-up":{"post":{"tags":["Authentications methods"],"summary":"Sign up user in the system (Client only!!)","description":"Sign up user in the systemm","produces":["application\/json"],"parameters":[{"name":"data","in":"body","description":"Data for sign up","required":true,"schema":{"properties":{"name":{"type":"string","default":""},"company":{"type":"string","default":""},"email":{"type":"string","default":""},"phone":{"type":"string","default":""},"password":{"type":"string","default":""},"password_confirmation":{"type":"string","default":""}},"type":"object"}}],"responses":{"200":{"description":"User's data"}}}},"\/client\/jobs":{"get":{"tags":["CLIENT:: Jobs methods"],"summary":"Get list of jobs with pagination","description":"Get list of jobs with pagination","produces":["application\/json"],"parameters":[{"name":"page","in":"query","required":true,"type":"integer","default":"1"}],"responses":{"200":{"description":"Jobs array"}},"security":[{"Bearer":[]}]},"post":{"tags":["CLIENT:: Jobs methods"],"summary":"Add new job","description":"Add new job","produces":["application\/json"],"parameters":[{"name":"data","in":"body","description":"Job's data","required":true,"schema":{"properties":{"client_id":{"type":"integer","default":""},"description":{"type":"string","default":""}},"type":"object"}}],"responses":{"200":{"description":"Job's data"}},"security":[{"Bearer":[]}]}},"\/client\/jobs\/{id}":{"get":{"tags":["CLIENT:: Jobs methods"],"summary":"View job data","description":"View job data","produces":["application\/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"integer"}],"responses":{"200":{"description":"Job data"}},"security":[{"Bearer":[]}]},"put":{"tags":["CLIENT:: Jobs methods"],"summary":"Edit the job","description":"Edit the job","produces":["application\/json"],"parameters":[{"name":"id","in":"path","description":"Number of job","required":true,"type":"integer","default":""},{"name":"data","in":"body","description":"Job's data","required":true,"schema":{"properties":{"client_id":{"type":"integer","default":""},"description":{"type":"string","default":""},"active":{"type":"integer","default":"0"},"completed":{"type":"integer","default":"0"}},"type":"object"}}],"responses":{"200":{"description":"Job's data"}},"security":[{"Bearer":[]}]}},"\/client\/reviews":{"get":{"tags":["CLIENT:: Reviews methods"],"summary":"Get list of client's review with pagination","description":"Get list of client's review with pagination","produces":["application\/json"],"parameters":[{"name":"page","in":"query","required":true,"type":"integer","default":"1"}],"responses":{"200":{"description":"Reviews array"}},"security":[{"Bearer":[]}]},"post":{"tags":["CLIENT:: Reviews methods"],"summary":"Add new review","description":"Add new revie","produces":["application\/json"],"parameters":[{"name":"data","in":"body","description":"Review's data","required":true,"schema":{"properties":{"client_id":{"type":"integer","default":""},"talent_id":{"type":"integer","default":""},"job_id":{"type":"integer","default":""},"rating":{"type":"integer","default":""},"text":{"type":"string","default":""}},"type":"object"}}],"responses":{"200":{"description":"New review data"}},"security":[{"Bearer":[]}]}},"\/client\/reviews\/{id}":{"get":{"tags":["CLIENT:: Reviews methods"],"summary":"View review data","description":"View review data","produces":["application\/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"integer"}],"responses":{"200":{"description":"Review data"}},"security":[{"Bearer":[]}]},"put":{"tags":["CLIENT:: Reviews methods"],"summary":"Edit review","description":"Edit review","produces":["application\/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"integer"},{"name":"data","in":"body","description":"Review's data","required":true,"schema":{"properties":{"client_id":{"type":"integer","default":""},"talent_id":{"type":"integer","default":""},"job_id":{"type":"integer","default":""},"rating":{"type":"integer","default":""},"text":{"type":"string","default":""}},"type":"object"}}],"responses":{"200":{"description":"Actually review data"}},"security":[{"Bearer":[]}]},"delete":{"tags":["CLIENT:: Reviews methods"],"summary":"Delete review","description":"Delete review","produces":["application\/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"integer"}],"responses":{"204":{"description":"Confirm"}},"security":[{"Bearer":[]}]}},"\/partner\/list":{"get":{"tags":["PUBLIC:: Partner's methods"],"summary":"Get list of talent partners","description":"Get list of talent partners","produces":["application\/json"],"responses":{"200":{"description":"Jobs data"}}}},"\/partner\/cv":{"post":{"tags":["PUBLIC:: Partner's methods"],"summary":"Send an application for partner-candidate","description":"","consumes":["multipart\/form-data"],"produces":["application\/json"],"parameters":[{"name":"resume","in":"formData","description":"Resume file","required":false,"type":"file"},{"name":"name","in":"formData","description":"Candidate's name","required":true,"type":"string"},{"name":"email","in":"formData","description":"Candidate's email","required":true,"type":"string"},{"name":"phone","in":"formData","description":"Candidate's phone","required":true,"type":"string"}],"responses":{"200":{"description":"Candidate's data"}}}},"\/profile":{"get":{"tags":["AUTHENTICATED USERS:: Profile methods"],"summary":"View profile of current user","description":"View profile of current user","produces":["application\/json"],"responses":{"200":{"description":"User's data"}},"security":[{"Bearer":[]}]},"delete":{"tags":["AUTHENTICATED USERS:: Profile methods"],"summary":"Drop user's record","description":"Drop user's record","produces":["application\/json"],"responses":{"204":{"description":"Confirmation of removal"}},"security":[{"Bearer":[]}]}},"\/profile\/{id}":{"post":{"tags":["AUTHENTICATED USERS:: Profile methods"],"summary":"Edit profile of current user","description":"Edit profile of current user","consumes":["multipart\/form-data"],"produces":["application\/json"],"parameters":[{"name":"avatar_image","in":"formData","description":"Avatar Image","required":false,"type":"file"},{"name":"name","in":"formData","description":"User name","required":true,"type":"string"},{"name":"avatar","in":"formData","description":"User avatar (link)","required":false,"type":"string"},{"name":"email","in":"formData","description":"User email","required":true,"type":"string"},{"name":"phone","in":"formData","description":"User phone","required":true,"type":"string"}],"responses":{"200":{"description":"User data"}}}},"\/seeker\/cv":{"post":{"tags":["PUBLIC:: Candidate's methods"],"summary":"Send an application for job seeker","description":"","consumes":["multipart\/form-data"],"produces":["application\/json"],"parameters":[{"name":"resume","in":"formData","description":"Resume file","required":false,"type":"file"},{"name":"name","in":"formData","description":"Candidate's name","required":true,"type":"string"},{"name":"email","in":"formData","description":"Candidate's email","required":true,"type":"string"},{"name":"phone","in":"formData","description":"Candidate's phone","required":true,"type":"string"}],"responses":{"200":{"description":"Candidate's data"}}}},"\/talent\/candidates":{"get":{"tags":["TALENT:: Candidate's methods"],"summary":"Get list of candidates with pagination","description":"Get list of candidates with pagination","produces":["application\/json"],"parameters":[{"name":"page","in":"query","description":"Number of page","required":true,"type":"integer","default":"1"}],"responses":{"200":{"description":"User's data"}},"security":[{"Bearer":[]}]}},"\/talent\/jobs":{"get":{"tags":["TALENT:: Job's methods"],"summary":"Get list of talent's jobs with pagination","description":"Get list of talent's jobs with pagination","produces":["application\/json"],"parameters":[{"name":"page","in":"query","description":"Number of page","required":true,"type":"integer","default":"1"}],"responses":{"200":{"description":"Jobs data"}},"security":[{"Bearer":[]}]}},"\/talent\/partner-candidates":{"get":{"tags":["TALENT:: Partner-candidate's methods"],"summary":"Get list of partner-candidates with pagination","description":"Get list of partner-candidates with pagination","produces":["application\/json"],"parameters":[{"name":"page","in":"query","description":"Number of page","required":true,"type":"integer","default":"1"}],"responses":{"200":{"description":"Candidate's data"}},"security":[{"Bearer":[]}]}},"\/talent\/reviews":{"get":{"tags":["TALENT:: Reviews methods"],"summary":"Get list of review with pagination","description":"Get list of review with pagination","produces":["application\/json"],"parameters":[{"name":"page","in":"query","required":true,"type":"integer","default":"1"}],"responses":{"200":{"description":"Reviews array"}},"security":[{"Bearer":[]}]}},"\/upload\/resume":{"post":{"tags":["UPLOAD:: Upload methods"],"summary":"","description":"","consumes":["multipart\/form-data"],"produces":["application\/json"],"parameters":[{"name":"resume","in":"formData","description":"file to upload","required":false,"type":"file"}],"responses":{"200":{"description":"User's data"}},"security":[{"Bearer":[]}]}}},"definitions":{},"securityDefinitions":{"Bearer":{"type":"apiKey","description":"qweqwe","name":"Authorization","in":"header"}}}